Rename file & remove other content.
